Fix remove_invisible not working
Prior to this, using {: or :} for brackets would not prune the invisible
elements from output. Since getchildren() is not a direct view on the
children of an element, deleting from ns[i].getchildren() has no effect.

In order to actually remove an XML element, the parent needs to be
directly passed. Then, the parent's remove method must be called since
that will look up the child by instance before deleting it.

Finally, make these test cases readable. It's hard to read the test
cases when they span so many columns.