Pass testBar
The code before was not backwards parsing for symmetrical parenthesis
like the comments said. This resulted in expressions being properly
bracketed whenever | or || was involved.

The fix for this comes in three parts:

- Properly test whether we are starting a nest.
  For symmetrical brackets, we must parse **backwards** to collect the
  appropriate contents into its own mrow.
  If not, then leave it alone.
- make find_node_backwards report the proper index since the previous
  result would always report the node ahead of the target.
- Properly test whether the descending expression actually finishes off
  our set of brackets.

The fix so far doesn't wrap the solo symmetrical brackets in their own <mrow>s.
Why? They're treated as infix expressions in those cases, so it makes no
sense for them to be treated as their own group.