Pass testNegative
Prior to this, all minuses were grouped with numbers unconditionally.
The following patch fixes this by checking if the preceding term is an
identifier. If so, put the number and the minus sign outside of the
element.

However, it still differentiates from the js implementation since
`-123/16` and `- 123/16` emit different XML.